add import Patients

+++ src/main/java/edu/pitt/medschool/service/ImportCsvService.java
@@ -788,7 +788,6 @@ public class ImportCsvService {
     }
 
     private void moveAndDelete(String path, Boolean success){
-        System.out.println(path);
         try {
             File afile = new File(path);
             String destnation;
@@ -805,10 +804,13 @@ public class ImportCsvService {
                 System.out.println("File "+afile.getName()+" is failed to move!");
             }
             String txtpath = path.replace(".csv",".txt");
-            System.out.println(txtpath);
             File txt = new File(txtpath);
             if(txt.exists()){
-                txt.delete();
+                if(success){
+                    txt.delete();
+                }else{
+                    moveFile(destnation,txt);
+                }
             }
         } catch (Exception e) {
             e.printStackTrace();
